`songrequest` - staticData refactor + other
- moved Static_Data partially into consts and `cytube-integration.js`
- removed blacklists
